+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clarified what `Config.Environment` affects.
 - AI Guard: replaced `redacted` with `action`.
 
+### Removed
+
+- Prompt Guard: `threshold` parameter.
+
 ## 4.4.0-beta.2 - 2025-01-21
 
 ### Added