+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - AI Guard: support for structured input.
 - AI Guard: more detectors.
+- AI Guard: support for `llm_input`, `llm_info`, and `log_fields`.
 
 ### Changed