+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Redact: `Unredact()` method on service interface.
+- Redact: `FPEContext` on `RedactResult` and `StructuredResult`.
+
 ## 4.4.0 - 2025-02-16
 
 ### Added