+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact: `Unredact()` method on service interface.
 - Redact: `FPEContext` on `RedactResult` and `StructuredResult`.
 
+### Removed
+
+- AI Guard: `LlmInfo` and `LlmInput`.
+
 ## 4.4.0 - 2025-02-16
 
 ### Added