+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Redact: `Unredact()` method on service interface.
 - Redact: `FPEContext` on `RedactResult` and `StructuredResult`.
+- PangeaConfig: `BaseURLTemplate` support.
 
 ### Changed
 
@@ -19,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - AI Guard: `LlmInfo` and `LlmInput`.
+- PangeaConfig: `Domain` and `Environment`. Use `BaseURLTemplate` instead.
 
 ## 4.4.0 - 2025-02-16