+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 5.0.0 - 2025-03-28
+
 ### Added
 
 - `BaseURLTemplate` has been added to `Config` to allow for greater control over